fix(slider): fix bugs with thumbs
Fixed issue with thumbs moving too fast which was caused by the `Track`
receiving pointer events. Fixed by adding `stopPropagation`.
Fixed issue with the wrong thumb being selected when thumb values are 0
and they overlap.
Changed onSlideStart to take in an index and a value. Thumbs pass in
their own index and track will use `getClosestValueIndex` method to
calculate the thumb that needs to be moved.